chore: Renamed handler folder to http. Refactored package names accordingly
